requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain>=0.1.0
langchain-community>=0.0.21
langchain-core>=0.1.27
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
chromadb==0.4.22
pypdf==3.17.1
python-multipart==0.0.6
numpy<2.0.0